# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación API ",
        "description": "Documentación API ",
        "contact": {
            "name": "API Support",
            "email": "tommycontreras34@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get list of users",
                "description": "Return list of users",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create a new user",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "username": {
                                                "type": "string"
                                            },
                                            "first_name": {
                                                "type": "string"
                                            },
                                            "last_name": {
                                                "type": "string"
                                            },
                                            "telephone": {
                                                "type": "string"
                                            },
                                            "age": {
                                                "type": "number"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "username": "Tommy 11",
                                    "first_name": "Tommy",
                                    "last_name": "Grullón Contreras",
                                    "telephone": "829-754-6150",
                                    "age": 20
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "username": {
                                            "type": "string",
                                            "example": "Tommy 11"
                                        },
                                        "first_name": {
                                            "type": "string",
                                            "example": "Tommy"
                                        },
                                        "last_name": {
                                            "type": "string",
                                            "example": "Grullón Contreras"
                                        },
                                        "telephone": {
                                            "type": "string",
                                            "example": "829-754-6150"
                                        },
                                        "age": {
                                            "type": "number",
                                            "example": 20
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Show user information",
                "operationId": "36a33ff774d5cba33c039dec2c3e0287",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "username": {
                                            "type": "string",
                                            "example": "Tommy 11"
                                        },
                                        "first_name": {
                                            "type": "string",
                                            "example": "Tommy"
                                        },
                                        "last_name": {
                                            "type": "string",
                                            "example": "Grullón Contreras"
                                        },
                                        "telephone": {
                                            "type": "string",
                                            "example": "829-754-6150"
                                        },
                                        "age": {
                                            "type": "integer",
                                            "example": "20"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete user information",
                "operationId": "fa56cffde745d3f152f95cbacd936c0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "NO CONTENT"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "User"
                ],
                "summary": "Update user information",
                "operationId": "872a537e8eb13c26c5542c7d73269a80",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "username": {
                                                "type": "string"
                                            },
                                            "first_name": {
                                                "type": "string"
                                            },
                                            "last_name": {
                                                "type": "string"
                                            },
                                            "telephone": {
                                                "type": "string"
                                            },
                                            "age": {
                                                "type": "number"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "username": "Tommy 11",
                                    "first_name": "Tommy",
                                    "last_name": "Grullón Contreras",
                                    "telephone": "829-754-6150",
                                    "age": 20
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "CREATED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "username": {
                                            "type": "string",
                                            "example": "Tommy 11"
                                        },
                                        "first_name": {
                                            "type": "string",
                                            "example": "Tommy"
                                        },
                                        "last_name": {
                                            "type": "string",
                                            "example": "Grullón Contreras"
                                        },
                                        "telephone": {
                                            "type": "string",
                                            "example": "829-754-6150"
                                        },
                                        "age": {
                                            "type": "number",
                                            "example": 20
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    }
}